Old Functions

The VDisp program has kept in some of the old functionality of VDispl, like parsing the old input file format and directly writing to an output file. The functions assisting with this process are still available for any future contributors or curious developers to play around with and are listed below:

vdisp.readInputFile — Function

readInputFile(inputPath, outputPath)

Reads and parses input file at inputPath and outputs calculations to file at outputPath.

This function was used to emulate old VDispl software's CLI funcitonality. It is no longer used in this version of VDisp. It has been left in for any developers that would like to have the command line funcitonality.