Output

VDisp output files are given a default format which include a repeat all the input data, and all relevant output data. The effective stress data are written to a file with the same name, but an extension of ".eff.dat", by the writeEffectiveStress function.

Main.OutputFormat.writeDefaultOutput — Function

writeDefaultOutput(outputData::OutputData, path::String)

Writes the outputs in the default format which includes all the input data for traceability purposes, and all the outputs after. By default, it also writes effective stress values to a file with the same name but different file extension (".eff.dat"). For more info on the effective stress outputs, see writeEffectiveStress docs.

Main.OutputFormat.writeEffectiveStresses — Function

writeEffectiveStresses(outputData, outputPath::String)

Writes table of effective stress values of each soil sublayer before and after applying the foundation load to file at outputPath.

VDisp calls this function from within the writeDefaultOutput() function. Thus, by default every VDisp output includes a separate file which displays the effective stress values. This file has the same filepath and name as the normal output file, but has the ".eff.dat" extension rather than ".dat".